Clean up reliable refund agent demo
