+++ examples/reliable_refund/agents.py
@@ -1,5 +1,5 @@
-from swarm import Agent
 from dbos import DBOS
+from swarm import Agent
 
 def process_refund(context_variables, item_id, reason="NOT SPECIFIED"):
     """Refund an item. Refund an item. Make sure you have the item_id of the form item_... Ask for user confirmation before processing the refund."""
@@ -21,26 +21,8 @@ def apply_discount():
     print("[mock] Applying discount...")
     return "Applied discount of 11%"
 
-triage_agent = Agent(
-    name="Triage Agent",
-    instructions="Determine which agent is best suited to handle the user's request, and transfer the conversation to that agent.",
-)
-
 refunds_agent = Agent(
     name="Refunds Agent",
     instructions="Help the user with a refund. If the reason is that it was too expensive, offer the user a refund code. If they insist, then process the refund.",
     functions=[process_refund, apply_discount],
 )
-
-
-def transfer_back_to_triage():
-    """Call this function if a user is asking about a topic that is not handled by the current agent."""
-    return triage_agent
-
-def transfer_to_refunds():
-    """Transfer the user to the refunds agent."""
-    return refunds_agent
-
-
-triage_agent.functions = [transfer_to_refunds]
-refunds_agent.functions.append(transfer_back_to_triage)
