--- examples/reliable_refund/durable_swarm.py
@@ -1,28 +1,0 @@
-from swarm import Swarm
-from dbos import DBOS, DBOSConfiguredInstance
-from swarm.repl.repl import pretty_print_messages
-
-DBOS()
-
-@DBOS.dbos_class()
-class DurableSwarm(Swarm, DBOSConfiguredInstance):
-    def __init__(self, client=None):
-        Swarm.__init__(self, client)
-        DBOSConfiguredInstance.__init__(self, "openai_client")
-
-    @DBOS.step()
-    def get_chat_completion(self, *args, **kwargs):
-        return super().get_chat_completion(*args, **kwargs)
-
-    # In this example, decorate individual tool calls with @DBOS.step() instead
-    def handle_tool_calls(self, *args, **kwargs):
-        return super().handle_tool_calls(*args, **kwargs)
-
-    @DBOS.workflow()
-    def run(self, *args, **kwargs):
-        response = super().run(*args, **kwargs)
-        pretty_print_messages(response.messages)
-        return response
-
-
-DBOS.launch()
