+++ examples/reliable_refund/main.py
@@ -1,13 +1,33 @@
-import sys
 import signal
+import sys
 
+from dbos import DBOS, DBOSConfiguredInstance, SetWorkflowID
+from swarm import Swarm
+from swarm.repl.repl import pretty_print_messages
 from agents import refunds_agent
-from durable_swarm import DurableSwarm
-from dbos import SetWorkflowID
 
-client = DurableSwarm()
+DBOS()
+
+@DBOS.dbos_class()
+class DurableSwarm(Swarm, DBOSConfiguredInstance):
+    def __init__(self, client=None):
+        Swarm.__init__(self, client)
+        DBOSConfiguredInstance.__init__(self, "openai_client")
+
+    @DBOS.step()
+    def get_chat_completion(self, *args, **kwargs):
+        return super().get_chat_completion(*args, **kwargs)
+
+    @DBOS.workflow()
+    def run(self, *args, **kwargs):
+        response = super().run(*args, **kwargs)
+        pretty_print_messages(response.messages)
+        return response
+
+DBOS.launch()
 
 def main():
+    client = DurableSwarm()
     print("Connecting to Durable Refund Agent 💪🐝")
 
     user_name = input("\033[90mWhat's your name\033[0m: \n")
@@ -18,7 +38,6 @@ def main():
     context_variables = {"user_name": user_name}
 
     # SetWorkflowID is used to ensure that user is refunded exactly once.
-    # If a workflow was interrupted and restarted, then it'll resume from where it left off.
     with SetWorkflowID(user_name):
         client.run(
             agent=refunds_agent,
